requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.20
python-dotenv==1.0.0
